+++ apis/alb/v1alpha1/dependencies.go
@@ -12,12 +12,12 @@ type DatacenterConfig struct {
 	// +crossplane:generate:reference:type=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.Datacenter
 	// +crossplane:generate:reference:extractor=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.ExtractDatacenterID()
 	DatacenterID string `json:"datacenterId,omitempty"`
-	// DatacenterIDRef references to a Datacenter to retrieve its ID
+	// DatacenterIDRef references to a Datacenter to retrieve its ID.
 	//
 	// +optional
 	// +immutable
 	DatacenterIDRef *xpv1.Reference `json:"datacenterIdRef,omitempty"`
-	// DatacenterIDSelector selects reference to a Datacenter to retrieve its datacenterId
+	// DatacenterIDSelector selects reference to a Datacenter to retrieve its DatacenterID.
 	//
 	// +optional
 	DatacenterIDSelector *xpv1.Selector `json:"datacenterIdSelector,omitempty"`
@@ -32,30 +32,32 @@ type LanConfig struct {
 	// +crossplane:generate:reference:type=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.Lan
 	// +crossplane:generate:reference:extractor=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.ExtractLanID()
 	LanID string `json:"lanId,omitempty"`
-	// LanIDRef references to a Lan to retrieve its ID
+	// LanIDRef references to a Lan to retrieve its ID.
 	//
 	// +optional
 	// +immutable
 	LanIDRef *xpv1.Reference `json:"lanIdRef,omitempty"`
-	// LanIDSelector selects reference to a Lan to retrieve its lanId
+	// LanIDSelector selects reference to a Lan to retrieve its LanID.
 	//
 	// +optional
 	LanIDSelector *xpv1.Selector `json:"lanIdSelector,omitempty"`
 }
 
-// IPsConfigs - used by resources that need to link multiple IPs from IPBlock via id or via reference
-// and using index. Indexes start from 0, and multiple indexes can be set.
-// If no index is set, all IPs from the corresponding IPBlock will be assigned.
-// If both IPs and IPBlockConfigs fields are set, only ips will be considered.
+// IPsConfigs - used by resources that need to link multiple IPs directly or from IPBlock via id or via reference.
 type IPsConfigs struct {
-	IPs         []string         `json:"ips,omitempty"`
+	// Use IPs to set specific IPs to the resource. If both IPs and IPsBlockConfigs are set,
+	// only `ips` field will be considered.
+	IPs []string `json:"ips,omitempty"`
+	// Use IpsBlockConfigs to reference existing IPBlocks, and to mention the indexes for the IPs.
+	// Indexes start from 0, and multiple indexes can be set. If no index is set, all IPs from the
+	// corresponding IPBlock will be assigned to the resource.
 	IPBlockCfgs []IPsBlockConfig `json:"ipsBlockConfigs,omitempty"`
 }
 
 // IPsBlockConfig - used by resources that need to link IPBlock via id or via reference
 // to get multiple IPs.
 type IPsBlockConfig struct {
-	// NicID is the ID of the IPBlock on which the resource will be created.
+	// IPBlockID is the ID of the IPBlock on which the resource will be created.
 	// It needs to be provided via directly or via reference.
 	//
 	// +immutable
@@ -63,12 +65,12 @@ type IPsBlockConfig struct {
 	// +crossplane:generate:reference:type=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.IPBlock
 	// +crossplane:generate:reference:extractor=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.ExtractIPBlockID()
 	IPBlockID string `json:"ipBlockId,omitempty"`
-	// IPBlockIDRef references to a IPBlock to retrieve its ID
+	// IPBlockIDRef references to a IPBlock to retrieve its ID.
 	//
 	// +optional
 	// +immutable
 	IPBlockIDRef *xpv1.Reference `json:"ipBlockIdRef,omitempty"`
-	// IPBlockIDSelector selects reference to a IPBlock to retrieve its nicId
+	// IPBlockIDSelector selects reference to a IPBlock to retrieve its IPBlockID.
 	//
 	// +optional
 	IPBlockIDSelector *xpv1.Selector `json:"ipBlockIdSelector,omitempty"`
@@ -80,19 +82,22 @@ type IPsBlockConfig struct {
 	Indexes []int `json:"indexes,omitempty"`
 }
 
-// IPConfig is used by resources that need to link ips from IPBlock via id or via reference
-// and using index. Indexes start from 0, and only one index must be set.
-// If both IPs and IPBlockConfigs fields are set, only ip will be used.
+// IPConfig is used by resources that need to link ip directly or from IPBlock via id or via reference.
 type IPConfig struct {
+	// Use IP to set specific IP to the resource. If both IP and IPBlockConfig are set,
+	// only `ip` field will be considered.
+	//
 	// +kubebuilder:validation:Pattern="^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?).){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
-	IP         string        `json:"ip,omitempty"`
+	IP string `json:"ip,omitempty"`
+	// Use IpBlockConfig to reference existing IPBlock, and to mention the index for the IP.
+	// Index starts from 0 and it must be provided.
 	IPBlockCfg IPBlockConfig `json:"ipBlockConfig,omitempty"`
 }
 
 // IPBlockConfig - used by resources that need to link IPBlock via id or via reference
 // to get one single IP.
 type IPBlockConfig struct {
-	// NicID is the ID of the IPBlock on which the resource will be created.
+	// IPBlockID is the ID of the IPBlock on which the resource will be created.
 	// It needs to be provided via directly or via reference.
 	//
 	// +immutable
@@ -100,17 +105,17 @@ type IPBlockConfig struct {
 	// +crossplane:generate:reference:type=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.IPBlock
 	// +crossplane:generate:reference:extractor=github.com/ionos-cloud/crossplane-provider-ionoscloud/apis/compute/v1alpha1.ExtractIPBlockID()
 	IPBlockID string `json:"ipBlockId,omitempty"`
-	// IPBlockIDRef references to a IPBlock to retrieve its ID
+	// IPBlockIDRef references to a IPBlock to retrieve its ID.
 	//
 	// +optional
 	// +immutable
 	IPBlockIDRef *xpv1.Reference `json:"ipBlockIdRef,omitempty"`
-	// IPBlockIDSelector selects reference to a IPBlock to retrieve its nicId
+	// IPBlockIDSelector selects reference to a IPBlock to retrieve its IPBlockID.
 	//
 	// +optional
 	IPBlockIDSelector *xpv1.Selector `json:"ipBlockIdSelector,omitempty"`
 	// Index is referring to the IP index retrieved from the IPBlock.
-	// Index is starting from 0.
+	// Index starts from 0.
 	//
 	// +kubebuilder:validation:Required
 	Index int `json:"index"`